reverse inf flag, tidying, apply black
