+++ cechmate/filtrations/__init__.py
@@ -1,3 +1,3 @@
-from .alpha import Alpha
-from .rips import Rips
-from .verenice import Verenice
+from .alpha import *
+from .rips import *
+from .cover import *
