+++ test/test_cover.py
@@ -1,6 +1,6 @@
 import pytest
 
-from cechmate import Verenice
+from cechmate import Cover
 
 
 @pytest.fixture
@@ -10,7 +10,7 @@ def two_cover():
 
 
 def test_single_edge(two_cover):
-    v = Verenice(2).build(two_cover)
+    v = Cover(2).build(two_cover)
 
     assert len(v) == 3
 
@@ -19,7 +19,7 @@ def test_jaccard_dist_equal():
     a = [0, 1]
     b = [0, 1]
 
-    v = Verenice(2)
+    v = Cover(2)
 
     assert v.jaccard([a, b]) == 0
 
@@ -29,7 +29,7 @@ def test_jaccard_dist_disjoint():
     b = [4, 5]
     c = [10, 20]
 
-    v = Verenice(2)
+    v = Cover(2)
 
     assert v.jaccard([a, b, c]) == 1
 
@@ -39,6 +39,6 @@ def test_triangles():
     b = [1, 5]
     c = [1, 6]
 
-    v = Verenice(3).build([a, b, c])
+    v = Cover(3).build([a, b, c])
 
     assert len(v) == 7
